fix(mac): add support for macOS and refactor nutrition target inputs
- Fixed compatibility issues on macOS, ensuring proper handling of TextField inputs for integers.
- Refactored macronutrient properties from Double to Integer for consistency and improved clarity.
- Created a standalone reusable component for displaying and managing each nutrition target input, simplifying the codebase and enhancing modularity.

+++ Logbog Diet/Features/Profile/Domain/Models/NutritionTarget.swift
@@ -8,9 +8,9 @@ import Foundation
 
 struct NutritionTarget: Equatable, Codable {
     var calories: Int         // Total daily calorie goal
-    var proteins: Double      // Protein goal in grams
-    var fats: Double          // Fat goal in grams
-    var carbs: Double         // Carbohydrate goal in grams
+    var proteins: Int      // Protein goal in grams
+    var fats: Int          // Fat goal in grams
+    var carbs: Int         // Carbohydrate goal in grams
 
     // Derived property: Calculates total macronutrient calories
     var totalMacroCalories: Int {
@@ -30,14 +30,14 @@ struct NutritionTarget: Equatable, Codable {
     func macroPercentages() -> (proteinPercent: Double, fatPercent: Double, carbPercent: Double)? {
         guard calories > 0 else { return nil }
         let totalCalories = Double(calories)
-        let proteinPercent = (proteins * 4) / totalCalories * 100
-        let fatPercent = (fats * 9) / totalCalories * 100
-        let carbPercent = (carbs * 4) / totalCalories * 100
+        let proteinPercent = (Double(proteins) * 4) / totalCalories * 100
+        let fatPercent = (Double(fats) * 9) / totalCalories * 100
+        let carbPercent = (Double(carbs) * 4) / totalCalories * 100
         return (proteinPercent, fatPercent, carbPercent)
     }
 
     // Factory method for a default target
     static func defaultTarget() -> NutritionTarget {
-        return NutritionTarget(calories: 2000, proteins: 100.0, fats: 70.0, carbs: 250.0)
+        return NutritionTarget(calories: 2000, proteins: 100, fats: 70, carbs: 250)
     }
 }
